update: Rename the macro name, Kohsaka Macro -> SIDAM
The prefix KM is still left almost everywhere at the moment. Hopefully this can be changed gradually.
Dependence to the RIKEN website is removed. Some help files are moved to local folders. The update check is stopped because it relies on the xml file of revision history on the web.
The version scheme is changed from a simple revision number to a sequence of three digits (major.minor.patch) because the version control system is changed from SVN to Git.