update: Several new color tables
The new color tables are those included in "Color tables" folder in the
Igor installation folder.

The color tables in Matplotlib1 and 2 are put in a single folder to
reduce number of tabs and the options are hidden to prevent the panel
becomes too large.

A utility function (xml2ibw) is saved in a separated file. This is used
when you make a color scale and not used for daily analysis.