update: Improve robustness of LineSpectra, LineProfile, and SpectrumViewer
The improved robustness is related to renaming and closing the panels.

Panels or target windows of LineSpectra/LineProfile/SpectrumViewer can be renamed.

Even if a panel of LineSpectra/LineProfile/SpectrumViewer(for MLS) is closed
before compiling the macro, the closing process runs after compiling the macro
so that the macro works without an error.

Backward compatibility is NOT supported for this modification. Thus, the panels
of LineSpectra/LineProfile/SpectrumViewer(for MLS) made by the previous versions
no longer work properly.

Together with this update, a small bug have been fixed.
The target windows are now correctly listed in the popupmenu of LineSpectra.