update: Fourier transform
The output mode of maginitude square gives a wave which is square of
a wave given by magnitude, as expected from its name. Because of this
change, a wave given by magnitude square gets 1/numpnts(w) of before.

The output mode of phase give a wave in radian rather than degree.

18 new window functions added in Igor Prof 8 are available.

In the panel, both a 1D profile curve and a 2D image of a window
function are shown without a tab.

The main function is renamed. No interface has been changed.
KMFFT() -> SIDAMFFT()