update: Write the include statement in each file in the func folder
Three core files, SIDAM_Menus, SIDAM_Constants, and SIDAM_Hook are
written in SIDAM_Procedures.ipf among the ipf files in the func folder.
Because of this change, even if files are removed from or added to the
func folder, it is no longer necessary to restart SIDAM for re-listing
ipf files.