add: Median subtraction
The code for subtracting plane, line, layer, and phase are rewritten
so that they work more efficiently and for any wave data type.

The ROI for plane subtraction is now available not only order=1 but also
the other orders.

The following function is replaced.
The interface has been slightly changed.
KMSubtraction() -> SIDAMSubtraction()