update: KMWavesStats is deprecated
If you want to keep using KMWaveStats() in your own macro, move
src/SIDAM/func/deprecated/KM WavesStats.ipf to src/SIDAM/extension.
Otherwise, the following commands are recommended.

To get a 1D average wave from a 3D wave (the resultant 1D wave is a
function of energy), use the following commmands.

  MatrixOP avgw = sumCols(sumRows(w3d)) / numpoints(w3d)
  Redimension/N=(numpnts(avgw)) avgw

If NaNs are contained in the 3D wave, use the following command instead
of the 1st line.

  MatrixOP/O avgw = sumCols(sumRows(replaceNaNs(w3d,0))) / sumCols(sumRows(greater(numtype(w3d)*(-1),-1)))

Similarly, use the following commands to get a standard deviation wave.
This works even if NaNs are contained in the 3D wave.

  MatrixOP sdevw = sqrt(varCols(redimension(w3d,numpoints(w3d),1)))
  Redimension/N=(numpnts(sdevw)) sdevw