fix: Change names of global variables and strings created for nanonis files
In Igor 8, liberal object names are allowed for global variables and
strings although it is not written in the manual. In Igor 9, they are
not allowed.

Names of global variables and strings for sxm files are changed so as to
satisfy the strict rule. Those for dat and 3ds files are changed as well
for Igor 9 and kept unchanged for Igor 8.