update: Remove the check of FFT availability in the menu
Previously, the FFT-related menu items (e.g., Fourier Transform) were
grayed out when FFT was not applicable to the target wave. This made the
menu a bit complicated because checking whether FFT is available may
take a long time. Now such menu items can be selected whether FFT is
applicable or not. When not applicable, the "Do it" button in the
subsequent panel is disabled.