fix: #19 and #20
Using the mouse wheel or pressing the page up/down buttons only changed
the plane of a displayed image and did not do anything else. i.e., the
change was not reflected to others such as the z-range adjustment, the
layer annotation and the controls of index and value in the control bar.
This occurred in Igor Pro 8 because the modification event of the hook
function was not called.