update: Load "save all" waves in a single wave
Many waves were created when a data file saved as the "save all" mode
was loaded. Now they are concatenated to one wave. A 2D wave is created
for a point spectroscopy, and a 4D wave is created for a grid
spectroscopy. The 2nd and 4th dimensions correspond to the sweeps.